product crud is working fine using the webpage. Changes: added new crud methods on EcommerceController and ProductService, added edit.html and updated list.html to show saved products data

+++ src/main/java/org/factoriaf5/ecommerce/service/ProductService.java
@@ -2,6 +2,7 @@ package org.factoriaf5.ecommerce.service;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.List;
 
 import org.factoriaf5.ecommerce.dto.ProductDTO;
 import org.factoriaf5.ecommerce.model.Product;
@@ -15,43 +16,65 @@ public class ProductService {
     @Autowired
     ProductRepository productRepository;
 
-    /* public List<Product>findAllProducts(){
+    public List<Product>findAllProducts(){
         return productRepository.findAll();
-    } */
+    }
 
     public Product findProductById(Long id){
         return productRepository.findById(id).orElse(null);
     }
 
     public void deleteProductById(Long id){
-        productRepository.deleteById(id);
+        Product productToDelete = findProductById(id);
+        if(productToDelete!=null){
+            String imgUrl = productToDelete.getImgUrl();
+            if(imgUrl!=null){
+                imageDeletionService(imgUrl);
+            }
+            productRepository.deleteById(id);
+        }
     }
 
-    public Product updateProduct(Product product){
-        return productRepository.save(product);
+    private void imageDeletionService(String imgUrl) {
+        File img = new File(imgUrl);
+        if(img.exists()){
+            img.delete();
+        }
     }
 
-    public Product createProduct(ProductDTO productDTO,MultipartFile img) throws IOException{
+    public Product saveProduct(ProductDTO productDTO,MultipartFile img) throws IOException{
         Product product = new Product();
+        product.setId(productDTO.getId());
         product.setDescription(productDTO.getDescription());
         product.setName(productDTO.getName());
         product.setPrice(productDTO.getPrice());
         product.setStock(productDTO.getStock());
-        product.setImgUrl(imageProcessingService(img));
+        if(!img.isEmpty()){
+            if(product.getId()!=null){
+                String url = findProductById(product.getId()).getImgUrl();
+                if(url!=null){
+                    imageDeletionService(url);
+                }
+            }
+            product.setImgUrl(imageCreationService(img));
+        }else if(product.getId()!=null){
+            product.setImgUrl(findProductById(product.getId()).getImgUrl());
+        }
         return productRepository.save(product);
     }
 
-    private String imageProcessingService(MultipartFile file) throws IOException{
+    private String imageCreationService(MultipartFile img) throws IOException{
         final String UPLOAD_DIR = System.getProperty("user.dir")+"/images";
         File uploadDir = new File(UPLOAD_DIR);
         System.out.println("ruta del directorio de carga de imagen -> "+uploadDir.getAbsolutePath());
         if(!uploadDir.exists()){
             uploadDir.mkdir();
         }
-        String filename = file.getOriginalFilename()+"_"+System.currentTimeMillis();
+        String filename = img.getOriginalFilename()+"_"+System.currentTimeMillis();
         uploadDir = new File(uploadDir,filename);
         System.out.println("ruta del archivo de imagen -> "+uploadDir.getAbsolutePath());
-        file.transferTo(uploadDir);
+        img.transferTo(uploadDir);
         return "images/"+filename;
     }
+
 }
